+++ source/DataFlowAnalysis.cpp
@@ -2,6 +2,7 @@
 #include "BitSliceList.hpp"
 #include "DriverMap.hpp"
 #include "NetlistBuilder.hpp"
+#include "OperationLowering.hpp"
 
 #include "slang/ast/Expression.h"
 #include "slang/ast/ValuePath.h"
@@ -224,6 +225,14 @@ void DataFlowAnalysis::handle(ast::ProceduralAssignStatement const &stmt) {
   }
 }
 
+void DataFlowAnalysis::visitRvalue(ast::Expression const &expr) {
+  if (builder.options.expandOperations) {
+    OperationLowering(*this).visitOperand(expr);
+  } else {
+    visit(expr);
+  }
+}
+
 void DataFlowAnalysis::handleAssignmentLegacy(
     ast::AssignmentExpression const &expr) {
   auto &node = builder.nodeFactory.createAssignment(expr);
@@ -242,7 +251,7 @@ void DataFlowAnalysis::handleAssignmentLegacy(
   }
 
   if (!expr.isLValueArg()) {
-    visit(expr.right());
+    visitRvalue(expr.right());
   }
 }
 
@@ -318,7 +327,7 @@ void DataFlowAnalysis::handle(ast::AssignmentExpression const &expr) {
         case BitSliceSource::Kind::Opaque: {
           auto savedLVal = isLValue;
           isLValue = false;
-          visit(*src.opaqueExpr);
+          visitRvalue(*src.opaqueExpr);
           isLValue = savedLVal;
           break;
         }